+++ src/apis/member/userop.py
@@ -234,8 +234,6 @@ async def post_member_resign(request: Request, response: Response, authorization
     try:
         if config.tracker.lower() == "tracksim":
             r = await arequests.delete(f"https://api.tracksim.app/v1/drivers/remove", data = {"steam_id": str(steamid)}, headers = {"Authorization": "Api-Key " + config.tracker_api_token}, dhrid = dhrid)
-        elif config.tracker.lower() == "navio":
-            r = await arequests.delete(f"https://api.navio.app/v1/drivers/{steamid}", headers = {"Authorization": "Bearer " + config.tracker_api_token}, dhrid = dhrid)
         if r.status_code == 401:
             tracker_app_error = f"{TRACKERAPP} API Error: Invalid API Token"
         elif r.status_code // 100 != 2: